refactor(workload-engine): split modules
